+++ lib/fs.js
@@ -10,7 +10,7 @@ import {
   createReadStream 
 } from "node:fs";
 import { parse, join } from "node:path";
-import { resolve, isRoot } from "./path.js";
+import { resolve } from "./path.js";
 import { createHash } from "node:crypto";
 import { isString } from "@xan105/is";
 import {
@@ -20,43 +20,37 @@ import {
   asBoolean
 } from "@xan105/is/opt";
 import { shouldStringNotEmpty } from "@xan105/is/assert";
-import { Failure, attempt } from "@xan105/error";
+import { attempt } from "@xan105/error";
+
+const mkdir = (dirPath) => fs.mkdir(dirPath, { recursive: true });
+const rm = (path) => fs.rm(path, { recursive: true, force: true });
 
 async function readFile(filePath, options){
-  let warn;
 
-  if (isString(options)) options = { encoding: options};
-  if (options){
-    warn = asBoolean(options.bomWarning) ?? true;
-    delete options.bomWarning;
-  }
+  if (isString(options)) options = { encoding: options };
   
   const data = await fs.readFile(filePath, options);
 
-  if (options && options.encoding &&
-      options.encoding.toLowerCase().startsWith("utf") && 
-      isString(data) &&
-      data.charCodeAt(0) === 0xFEFF) //BOM
-  {
-    if (warn) console.warn(`Stripped UTF BOM from "${filePath}"; Use options: {encoding: "utf...", bom: true} with writeFile() to restore it.`);
+  if (
+    options?.encoding?.toLowerCase().startsWith("utf") && 
+    isString(data) && data.charCodeAt(0) === 0xFEFF //BOM
+  ) 
     return data.slice(1);
-  }
-  else return data;
+  else 
+    return data;
 }
 
 async function readJSON(filePath){
-  const data = await readFile(filePath, { encoding: "utf8", bomWarning: false });
+  const data = await readFile(filePath, { encoding: "utf8" });
   return JSON.parse(data);
 }
 
-const mkdir = (dirPath) => fs.mkdir(dirPath, { recursive: true });
-
 async function writeFile(filePath, data, options){
   
-  if (isString(options)) options = { encoding: options};
+  if (isString(options)) options = { encoding: options };
   if (options){
     options.encoding ||= "utf8";
-    if (options.bom === true && options.encoding.toLowerCase().startsWith("utf") && isString(data)) data = "\ufeff" + data;
+    if (options.bom === true && options.encoding?.toLowerCase().startsWith("utf") && isString(data)) data = "\ufeff" + data;
     delete options.bom;
   }
 
@@ -138,20 +132,15 @@ async function stat(path) {
   return stats;
 }
 
-async function rm(path){
-  const [, err] = await attempt(fs.rm, [path]);
-  if (err?.code === "ENOTEMPTY"){
-    const files = await fs.readdir(path);
-    for (const { name } of files) await rm(join(path, name));
-    await attempt(fs.rm, [path]);
-  }
-}
-
-function hashFile(filePath, algo = "sha1"){
-  let sum = createHash(algo);
-  let stream = createReadStream(filePath);
-  
+function hashFile(filePath, algo = "sha256"){
   return new Promise((resolve, reject) => {
+  
+    shouldStringNotEmpty(filePath);
+    shouldStringNotEmpty(algo);
+    
+    const sum = createHash(algo);
+    const stream = createReadStream(filePath);
+    
     stream
     .on('error', (err) => {
       return reject(err);
@@ -196,7 +185,7 @@ async function ls(dirPath, option = {}) {
   
   let result = [];
   
-  const [ files = [] ] = await attempt(fs.readdir,[dirPath,{ 
+  const [ files = [] ] = await attempt(fs.readdir, [dirPath,{ 
     withFileTypes: true 
   }]);
   
@@ -226,9 +215,13 @@ async function ls(dirPath, option = {}) {
 }
 
 export {
+  mkdir,
+  rm,
+  rm as unlink, //alias (backward compatibility)
+  rm as deleteFile, //alias (backward compatibility)
+  rm as rmdir, //alias (backward compatibility)
   readFile,
   readJSON,
-  mkdir,
   writeFile,
   writeJSON,
   copyFile,
@@ -239,10 +232,6 @@ export {
   existsAndIsYoungerThan,
   stat,
   stat as stats, //alias (backward compatibility)
-  rm,
-  rm as unlink, //alias (backward compatibility)
-  rm as deleteFile, //alias (backward compatibility)
-  rm as rmdir, //alias (backward compatibility)
   hashFile,
   touch,
   ls
